move schedulingstate to job package
