+++ master/internal/job/job.go
@@ -4,7 +4,6 @@ import (
 	"errors"
 	"fmt"
 
-	"github.com/determined-ai/determined/master/internal/sproto"
 	"github.com/determined-ai/determined/master/pkg/actor"
 	"github.com/determined-ai/determined/master/pkg/model"
 	"github.com/determined-ai/determined/proto/pkg/apiv1"
@@ -54,7 +53,7 @@ func JobActorAddr(jobType model.JobType, entityId string) actor.Address {
 // RMJobInfo packs information available only to the RM that updates frequently.
 type RMJobInfo struct {
 	JobsAhead      int
-	State          sproto.SchedulingState
+	State          SchedulingState
 	RequestedSlots int
 	AllocatedSlots int
 	IsPreemptible  bool
